requirements.txt
streamlit==1.28.0
pandas==2.0.3
numpy==1.24.3
plotly==5.15.0
seaborn==0.12.2
matplotlib==3.7.2
ydata-profiling==4.6.0
scikit-learn==1.3.0
scipy==1.11.1
python-dotenv==1.0.0
google-generativeai==0.3.2
langchain==0.0.350
langchain-google-genai==0.0.5
sqlalchemy==2.0.20
pdfkit==1.0.0
fpdf==1.7.2
python-docx==0.8.11
kaleido==0.2.1